# backend/greenride-api-clean/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenRide API Documentation\n## Response Codes:\n- 0000: Success - 操作成功\n- 9999: System Error - 系统错误\n- 1000: Parameter Error - 参数错误\n- 1001: Authorization Error - 授权失败\n- 1002: Authentication Expired - 认证过期\n- 2001: User Error - 用户相关错误\n- 4001: Business Error - 业务错误\n- 4002: Data Expired - 数据过期",
        "title": "GreenRide API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/change-password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员修改自己的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-认证"
                ],
                "summary": "管理员修改密码",
                "parameters": [
                    {
                        "description": "密码修改信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.AdminChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/admin/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前登录管理员的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-认证"
                ],
                "summary": "获取当前管理员信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/protocol.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Admin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "管理员用户名密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.AdminLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/protocol.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Admin"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员退出登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-认证"
                ],
                "summary": "管理员登出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/admin/reset-password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重置指定管理员的密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-管理"
                ],
                "summary": "重置管理员密码",
                "parameters": [
                    {
                        "description": "密码重置信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.AdminResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户修改密码，需要提供旧密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "用户登录接口，支持邮箱和手机号登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误或认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出，清除服务端相关信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/reset-password": {
            "post": {
                "description": "通过验证码重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重置失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "user_type"
            ],
            "properties": {
                "email": {
                    "description": "login_type=email时必填",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "description": "login_type=phone时必填",
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_type": {
                    "description": "必填：passenger, driver，用于区分同一手机号/邮箱的不同身份",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                },
                "verify_code": {
                    "description": "login_type=phone时必填",
                    "type": "string"
                }
            }
        },
...
        "handlers.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "code",
                "method",
                "new_password",
                "user_type"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "method": {
                    "description": "email, phone",
                    "type": "string",
                    "enum": [
                        "email",
                        "phone"
                    ]
                },
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user_type": {
                    "description": "用户类型",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                }
            }
        },
...
        "models.Admin": {
            "type": "object",
            "properties": {
                "actions_this_month": {
                    "type": "integer"
                },
                "actions_this_week": {
                    "type": "integer"
                },
                "actions_today": {
                    "description": "统计数据",
                    "type": "integer"
                },
                "active_status": {
                    "description": "online, offline, busy",
                    "type": "string"
                },
                "admin_id": {
                    "type": "string"
                },
                "allowed_ips": {
                    "description": "允许的IP地址列表",
                    "type": "string"
                },
                "approved_at": {
                    "type": "integer"
                },
                "approved_by": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "description": "审计信息",
                    "type": "string"
                },
                "current_session_id": {
                    "description": "会话管理",
                    "type": "string"
                },
                "date_format": {
                    "type": "string"
                },
                "department": {
                    "description": "部门",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "emergency_contact": {
                    "type": "string"
                },
                "failed_attempts": {
                    "type": "integer"
                },
                "first_name": {
                    "description": "个人信息",
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_title": {
                    "description": "职位",
                    "type": "string"
                },
                "language": {
                    "description": "偏好设置",
                    "type": "string"
                },
                "last_failed_at": {
                    "type": "integer"
                },
                "last_login_at": {
                    "description": "登录相关",
                    "type": "integer"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "last_updated_by": {
                    "type": "string"
                },
                "locked_until": {
                    "type": "integer"
                },
                "login_count": {
                    "type": "integer"
                },
                "max_concurrent_sessions": {
                    "type": "integer"
                },
                "metadata": {
                    "description": "元数据",
                    "type": "string"
                },
                "must_change_password": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "string"
                },
                "office_address": {
                    "type": "string"
                },
                "office_phone": {
                    "description": "联系信息",
                    "type": "string"
                },
                "password_changed_at": {
                    "description": "安全设置",
                    "type": "integer"
                },
                "permissions": {
                    "description": "JSON数组存储权限列表",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "phone_verified": {
                    "type": "boolean"
                },
                "preferences": {
                    "description": "JSON对象存储个人偏好",
                    "type": "string"
                },
                "role": {
                    "description": "角色和权限",
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                },
                "security_question": {
                    "type": "string"
                },
                "session_count": {
                    "type": "integer"
                },
                "status": {
                    "description": "状态管理",
                    "type": "string"
                },
                "suspended_at": {
                    "type": "integer"
                },
                "suspended_by": {
                    "type": "string"
                },
                "suspension_reason": {
                    "type": "string"
                },
                "tags": {
                    "description": "逗号分隔的标签",
                    "type": "string"
                },
                "time_format": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "total_actions": {
                    "type": "integer"
                },
                "two_factor_enabled": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "integer"
                },
                "username": {
                    "description": "基本信息",
                    "type": "string"
                },
                "working_areas": {
                    "description": "JSON数组存储负责区域",
                    "type": "string"
                },
                "working_hours": {
                    "description": "工作时间和区域",
                    "type": "string"
                },
                "working_timezone": {
                    "type": "string"
                }
            }
        },
...
        "protocol.AdminChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
...
        "protocol.AdminLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "protocol.AdminResetPasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "target_admin_id"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "target_admin_id": {
                    "description": "目标管理员ID",
                    "type": "string"
                }
            }
        },
...
        "protocol.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}